Fix message flag update fetching

+++ src/MailCoreInterface/imap_idle_listener.cpp
@@ -30,7 +30,7 @@ private:
      GTask* task;
 };
 
- extern "C" void mail_core_interface_imap_idle_listener (mailcore::IMAPAsyncSession* session, gchar* folder_path, guint last_known_id, GAsyncReadyCallback callback, void* user_data) {
+ extern "C" void mail_core_interface_imap_idle_listener (mailcore::IMAPAsyncSession* session, gchar* folder_path, guint64 last_known_id, GAsyncReadyCallback callback, void* user_data) {
     auto task = g_task_new (NULL, NULL, callback, user_data);
 
     auto idle_operation = session->idleOperation (new mailcore::String(folder_path), last_known_id);
